requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
azure-common==1.1.28
azure-core==1.22.1
azure-identity==1.7.1
azure-keyvault-secrets==4.3.0
beautifulsoup4==4.12.2
bottle==0.12.23
bottle-websocket==0.2.9
bs4==0.0.1
cairocffi==1.4.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cryptography==36.0.1
cssselect2==0.7.0
cx-Oracle==8.3.0
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
Eel==0.15.1
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
iconify==0.0.103
idna==2.10
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
kids.cache==0.0.7
kiwisolver==1.4.4
libsass==0.21.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mock==4.0.3
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
multidict==6.0.4
numpy==1.22.2
oauthlib==3.2.0
openai==0.27.4
openpyxl==3.0.9
packaging==22.0
pandas==1.4.1
Pillow==9.3.0
platformdirs==2.5.2
portalocker==2.3.2
psycopg2==2.9.5
pycparser==2.21
pygame==2.4.0
PyJWT==2.3.0
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5Custom==1.0.1
PySide2==5.15.2.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==304
QT-PyQt-PySide-Custom-Widgets==0.6.3
qtsass==0.3.2
requests==2.23.0
requests-oauthlib==1.3.1
rollbar==0.16.3
shiboken2==5.15.2.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tinycss2==1.2.1
tk==0.1.0
tqdm==4.65.0
typing==3.7.4
tzdata==2022.7
urllib3==1.25.11
virtualenv==20.16.5
virtualenvwrapper-win==1.2.7
webencodings==0.5.1
Werkzeug==2.2.2
whichcraft==0.6.1
whiptail==0.2
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
zope.event==4.6
zope.interface==5.5.2
